# lib/api-zod/src/generated/types/noteDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for الأستاذ عباس علي الغالبي website
 * OpenAPI spec version: 0.1.0
 */

export interface NoteDetail {
  id: number;
  title: string;
  teacherName: string;
  /** @nullable */
  version?: string | null;
  /** @nullable */
  pageCount?: number | null;
  /** @nullable */
  fileSize?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
  /** @nullable */
  telegramDownloadUrl?: string | null;
  /** @nullable */
  telegramPurchaseUrl?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  averageRating?: number | null;
  ratingCount: number;
  viewCount: number;
  downloadClickCount: number;
  previewClickCount: number;
  updatedAt: Date;
  createdAt: Date;
}